podvm-mkosi: print info and PCRs to serial console
Working with the new images, two information about the image are of
fundamental interest both for devs as for users: the vTPM PCR values
of the image and whether the image is a debug image or intended for
production use. This changes adds printing capabilities for those
information.

As production images don't allow access via SSH or serial console, we
print everything to the serial console after boot. For dev convenience,
the debug image will also print the image info on SSH login.