podvm: remove mount units
Previously we had /peerpod and /kata-containers bind-mounted to /run
to keep the size of the tempFS under /run small. As / is now read-only
this does not make sense anymore, so removing it.

This triggered a bug in kata-agent/ttrpc-rust as the path of the socket
is not created automatically. Adding a workaround to the kata-agent
unit until this is fixed upstream.